feat: separate extensions and its usage
