+++ src/assets/js/extensions/ckeditor.js
@@ -1,7 +1,1 @@
 import ClassicEditor from "@ckeditor/ckeditor5-build-classic"
-
-ClassicEditor
-.create(document.querySelector('#editor'))
-.catch(error => {
-    console.error(error);
-});
