+++ src/assets/js/extensions/jquery.js
@@ -1,0 +1,3 @@
+let $  = require( 'jquery' );
+window.$ = $
+window.jQuery = $
