+++ src/assets/js/extensions/summernote.js
@@ -1,21 +1,1 @@
-$ = require("jquery");
 require("summernote/dist/summernote-lite");
-
-$('#summernote').summernote({
-    tabsize: 2,
-    height: 120,
-})
-$("#hint").summernote({
-    height: 100,
-    toolbar: false,
-    placeholder: 'type with apple, orange, watermelon and lemon',
-    hint: {
-        words: ['apple', 'orange', 'watermelon', 'lemon'],
-        match: /\b(\w{1,})$/,
-        search: function (keyword, callback) {
-            callback($.grep(this.words, function (item) {
-                return item.indexOf(keyword) === 0;
-            }));
-        }
-    }
-});
