+++ src/assets/js/extensions/ui-todolist.js
@@ -1,4 +1,1 @@
 import dragula from "dragula"
-
-window.dragula = dragula;
-dragula([document.getElementById("widget-todo-list")], { moves: function (e, a, t) { return t.classList.contains("cursor-move") } })
