+++ src/assets/js/pages/ckeditor.js
@@ -1,0 +1,6 @@
+
+ClassicEditor
+.create(document.querySelector('#editor'))
+.catch(error => {
+    console.error(error);
+});
