+++ src/assets/js/pages/dashboard.js
@@ -1,5 +1,3 @@
-import ApexCharts from "apexcharts";
-
 var optionsProfileVisit = {
 	annotations: {
 		position: 'back'
