+++ src/assets/js/pages/datatables.js
@@ -1,0 +1,2 @@
+let jquery_datatable = $("#table1").DataTable();
+console.log($("#table1"))
