+++ src/assets/js/pages/jquery.js
@@ -1,0 +1,1 @@
+let $  = require( 'jquery' );
