+++ src/assets/js/pages/ui-todolist.js
@@ -1,0 +1,2 @@
+
+dragula([document.getElementById("widget-todo-list")], { moves: function (e, a, t) { return t.classList.contains("cursor-move") } })
